% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Default combine function for lowerGPreff_massfun objects}
\usage{
add(delay_massfun1, delay_massfun2, ...)
}
\arguments{
\item{delay_massfun1}{first massfun object}

\item{delay_massfun2}{second massfun object}

\item{...}{extra args}
}
\value{
combined lowerGPreff massfuns
}
\description{
Default combine function for lowerGPreff_massfun objects
}
