% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_reff.R
\name{compute_reff}
\alias{compute_reff}
\title{Compute R effective from model outputs and generation interval}
\usage{
compute_reff(fit_output, gi)
}
\arguments{
\item{fit_output}{model fit}

\item{gi}{generation interval distribution}
}
\value{
Estimated R effective
}
\description{
Compute R effective from model outputs and generation interval
}
