% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_delays.R
\name{construct_delays}
\alias{construct_delays}
\title{Construct delays}
\usage{
construct_delays(
  ecdf1,
  ecdf2 = NULL,
  delay_range = c(-3, 28),
  output = c("probability", "cumulative density"),
  stepfun_output = FALSE
)
}
\arguments{
\item{stepfun_output}{}
}
\description{
Given a date range to compute delays over, and an empirical cdf, compute the probability of delay
having that length. Outputs either the probability or the cumulative density, as either a data
frame lined up against days-of-day or as a step function in the style of R native ecdf function.
Optionally, two cdfs can be used as input, in which case the probability and cdf of the additive
combined delay are calculated.
}
