% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epiwave_timeseries.R
\name{create_epiwave_greta_timeseries}
\alias{create_epiwave_greta_timeseries}
\title{Create a greta-compatible timeseries object}
\usage{
create_epiwave_greta_timeseries(dates, car, chr_prior)
}
\arguments{
\item{dates}{infection dates sequence}

\item{car}{case ascertainment rate; either a numeric value or an
\code{epiwave_timeseries} object}

\item{chr_prior}{a greta array representing the prior distribution for
the case hospitalisation rate}
}
\value{
a list with class \code{epiwave_greta_timeseries} containing
components \code{timeseries} (a tibble of dates) and \code{ihr} (a greta array of
implied hospitalisation rates)
}
\description{
The epiwave model functions expect data in a long format,
structured to have a value for every date. This function creates a
greta-compatible timeseries from a case ascertainment rate and a
hospitalisation rate prior.
}
