% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epiwave_timeseries.R
\name{create_epiwave_massfun_timeseries}
\alias{create_epiwave_massfun_timeseries}
\title{Expand a discrete PMF into a long tibble}
\usage{
create_epiwave_massfun_timeseries(dates, jurisdictions, value)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{value}{a \code{discrete_pmf} or \code{discrete_pmf_series} object to be
replicated in each cell}
}
\value{
a long tibble with the distribution replicated across all
date and jurisdiction combinations, with class
\code{epiwave_massfun_timeseries}
}
\description{
The epiwave model functions expect data in a long format,
structured to have a value for every unique date and jurisdiction pair.
This function creates a tibble of this structure from a single
\code{discrete_pmf} or \code{discrete_pmf_series} object that is replicated in
each cell.
}
