% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ihr_prior.R
\name{create_ihr_prior}
\alias{create_ihr_prior}
\title{Create infection hospitalisation rate prior}
\usage{
create_ihr_prior(car)
}
\arguments{
\item{car}{case ascertainment rate}
}
\value{
infection fatality rate
}
\description{
Create a list of the same structure as case ascertainment rate.
Generate an infection hospitalisation rate as a greta array.
}
