% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_timeseries.R
\name{create_lowerGPreff_dist_timeseries}
\alias{create_lowerGPreff_dist_timeseries}
\title{Expand distribution into long tibble}
\usage{
create_lowerGPreff_dist_timeseries(dates, jurisdictions, value)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{value}{distribution to be replicated in each cell}
}
\value{
long tibble with distribution replicated
}
\description{
The lowerGPreff model functions expect data in a long format,
which is structured to have a value for every unique date and jurisdiction
pair. This function create a tibble of this structure out of a single
distribution that should be replicated in each cell.
}
