% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_timeseries.R
\name{create_lowerGPreff_fixed_timeseries}
\alias{create_lowerGPreff_fixed_timeseries}
\title{Title}
\usage{
create_lowerGPreff_fixed_timeseries(dates, jurisdictions, value)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{value}{constant value to be replicated in each cell}
}
\value{
long tibble with fixed value replicated
}
\description{
Title
}
