% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_massfun.R
\name{create_lowerGPreff_massfun}
\alias{create_lowerGPreff_massfun}
\title{Create lowerGPreff_massfun object}
\usage{
create_lowerGPreff_massfun(
  min_delay,
  max_delay,
  cdf_fun,
  normalise = c(TRUE, FALSE)
)
}
\arguments{
\item{min_delay}{miniumum delay}

\item{max_delay}{maxiumum delay}

\item{cdf_fun}{function for creating the cdf that defines the mass}

\item{normalise}{whether the mass should be normalised}
}
\value{
lowerGPreff_massfun object
}
\description{
Create lowerGPreff_massfun object
}
