% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_timeseries.R
\name{create_lowerGPreff_timeseries}
\alias{create_lowerGPreff_timeseries}
\title{Expand constant value into long tibble}
\usage{
create_lowerGPreff_timeseries(dates, jurisdictions, value)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{value}{value to be replicated in each cell}
}
\value{
long tibble with constant value replicated
}
\description{
The lowerGPreff model functions expect data in a long format,
which is structed to have a value for every unique date and jurisdiction
pair. This function create a tibble of this structure out of a single
value that should be replicated in each cell.
}
