% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observation_model.R
\name{create_observation_model}
\alias{create_observation_model}
\title{Create observation model}
\usage{
create_observation_model(
  data_id = "cases",
  observation_model_data,
  infection_days,
  infection_timeseries
)
}
\arguments{
\item{data_id}{name of observation model data in list}

\item{observation_model_data}{list of observation model data lists}

\item{infection_days}{infection dates that cover more than the data dates}

\item{infection_model}{greta arrays that define the infection model}
}
\value{
greta arrays of observation model
}
\description{
This function constructs the observation model and defines
likelihood over the observational data (e.g. cases, hospital
admissions...). It first begets a timeseries of expected observations from
a timeseries of new infections, the infection-to-observation delay
distribution(s), and the proportion of infections to be observed.
Specifically, the model uses the delay distribution(s) to convolve the
infection timeseries into the expected observation timeseries, thus
accounting for the probability of observation over time-since-infection. An
additional multiplication by observation proportion is applied to the
convolved timeseries, to adjust for effects such as case ascertainment.
Optionally, a day-of-week effect may be included in the convolution process
to introduce weekly periodicity. The expected observation timeseries is
treated as the mean of a negative binomial distribution from which the data
is observed, thus allowing us to define likelihood over the data, and link
the data to the unknown infection timeseries.
}
