% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_class_objs.R
\name{data_to_curve}
\alias{data_to_curve}
\title{CHANGE ME Create distribution object from parametric distribution values}
\usage{
data_to_curve(day_diff_col, value_col)
}
\arguments{
\item{day_diff_col}{days since onset of infection}

\item{value_col}{value on that many days after infection onset}
}
\value{
lowerGPreff_distribution object
}
\description{
For sero data the value col will be probability of testing
positive. For wastewater data the value col will be amount of viral
material shed by individual compared to peak amount.
}
