% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_to_matrix}
\alias{data_to_matrix}
\title{Turn wide data to matrix}
\usage{
data_to_matrix(long_data, extra_col)
}
\arguments{
\item{long_data}{long data form}

\item{extra_col}{for column name}
}
\value{
data in wide matrix form, with continuous seq of dates
}
\description{
Take long form input data and turn into wide form matrix with
continuous sequence of dates. This function assumes that users supply long
form count data where dates with 0 cases are explicit.
}
\keyword{internal}
