% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_observation_data.R
\name{define_sero_data}
\alias{define_sero_data}
\title{Define seroprevalence observation data}
\usage{
define_sero_data(
  timeseries_data,
  total_pop,
  size_vec,
  delay_from_infection,
  proportion_infections,
  dow_model = FALSE
)
}
\arguments{
\item{timeseries_data}{seroprevalence survey timeseries data, for one
jurisdiction}

\item{total_pop}{population size of this jurisdiction (a single numeric
value)}

\item{size_vec}{sample size of the seroprevalence survey, for one
jurisdiction}

\item{delay_from_infection}{delay data}

\item{proportion_infections}{proportion data}

\item{dow_model}{logical indicating whether to apply a DOW}
}
\value{
list of observation data for one data type
}
\description{
Bundle one jurisdiction's seroprevalence survey data. Call
this once per jurisdiction; combine multiple jurisdictions later via
\code{stack_jurisdictions()}.
}
