% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwave-package.R
\docType{package}
\name{epiwave-package}
\alias{epiwave}
\alias{epiwave-package}
\title{epiwave: Estimate epidemiological parameters by modelling new infections as a gaussian process over time}
\description{
This package exists as a toolbox to create the greta arrays necessary to model infections as a gaussian process over time. The resulting infection timeseries can be used to estimate epidemiological parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/idem-lab/epiwave}
  \item Report bugs at \url{https://github.com/idem-lab/epiwave/issues}
}

}
\author{
\strong{Maintainer}: Saras Windecker \email{saras.windecker@gmail.com} (\href{https://orcid.org/0000-0002-4870-8353}{ORCID})

Authors:
\itemize{
  \item Tianxiao Hao (\href{https://orcid.org/0000-0003-4363-1956}{ORCID})
  \item Nick Golding (\href{https://orcid.org/0000-0001-8916-5570}{ORCID})
}

Other contributors:
\itemize{
  \item Kate Senior (\href{https://orcid.org/0000-0003-1123-1472}{ORCID}) [contributor]
}

}
\keyword{internal}
