% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_reff.R
\name{estimate_reff}
\alias{estimate_reff}
\title{Estimate R_effective}
\usage{
estimate_reff(infection_timeseries, generation_interval_mass_fxns)
}
\arguments{
\item{generation_interval_mass_fxns}{}
}
\value{
greta arrays for infectiousness and $R_eff$
}
\description{
A short description...
}
