% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate the mass for given delay}
\usage{
evaluate(lookup, day_diff, ...)
}
\arguments{
\item{lookup}{lowerGPreff_distribution object with mass functions}

\item{day_diff}{matrix of}

\item{...}{extra args}
}
\value{
matrix of delay days with mass vals filled in
}
\description{
Evaluate the mass for given delay
}
