% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_delay_data.R
\name{extend_delay_data}
\alias{extend_delay_data}
\title{Title}
\usage{
extend_delay_data(delay_dat, incubation_period_distribution)
}
\arguments{
\item{incubation_period_distribution}{}
}
\description{
Title
}
