% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infection_timeseries.R
\name{gp_growth_rate}
\alias{gp_growth_rate}
\title{Effect type "growth_rate" formula for infection timeseries}
\usage{
gp_growth_rate(inits, z)
}
\arguments{
\item{inits}{prior for initial num. of infections on log scale}

\item{z}{gaussian process}
}
\value{
infection timeseries
}
\description{
Effect type "growth_rate" formula for infection timeseries
}
