% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infection_timeseries.R
\name{growth_rate_deriv}
\alias{growth_rate_deriv}
\title{Effect type "growth_rate_deriv" formula for infection timeseries}
\usage{
growth_rate_deriv(inits, z)
}
\arguments{
\item{inits}{prior for initial num. of infections on log scale}

\item{z}{gaussian process}
}
\value{
infection timeseries
}
\description{
Effect type "growth_rate_deriv" formula for infection timeseries
}
