% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_day_of_week.R
\name{implement_day_of_week}
\alias{implement_day_of_week}
\title{Implement day of week effect in observation model}
\usage{
implement_day_of_week(infection_days, dow_model)
}
\arguments{
\item{infection_days}{vector of dates that cover infection timeseries}

\item{dow_model}{greta model for day of week effect}
}
\value{
matrix of dimension date x jurisdiction with day of week correction
}
\description{
Implement day of week effect in observation model
}
