% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infection_timeseries.R
\name{infections}
\alias{infections}
\title{Effect type "infections" formula for infection timeseries}
\usage{
infections(inits, z)
}
\arguments{
\item{inits}{prior for initial num. of infections on log scale}

\item{z}{gaussian process}
}
\value{
infection timeseries
}
\description{
Effect type "infections" formula for infection timeseries
}
