% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits_by_jurisdiction.R
\name{inits_by_jurisdiction}
\alias{inits_by_jurisdiction}
\title{Define initial values for the infection timeseries}
\usage{
inits_by_jurisdiction(obs_data, delays, obs_prop, target_infection_dates)
}
\arguments{
\item{obs_data}{numeric vector of observed data, aligned to
\code{target_infection_dates} (\code{NA} for dates without observations)}

\item{delays}{epiwave massfun timeseries delay object, aligned to
\code{target_infection_dates}}

\item{obs_prop}{numeric vector of proportions, aligned to
\code{target_infection_dates}, or a greta array of the same}

\item{target_infection_dates}{infection date sequence}
}
\value{
initial values as a list
}
\description{
Define initial values for the infection timeseries
}
