% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerGPreff-package.R
\docType{package}
\name{lowerGPreff-package}
\alias{lowerGPreff}
\alias{lowerGPreff-package}
\title{lowerGPreff: Estimate epidemiological parameters by modelling new infections as a gaussian process over time}
\description{
What the package does (one paragraph). non discrete renewal. infection as non param of time.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
