% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cdf.R
\name{make_cdf}
\alias{make_cdf}
\title{Make cumulative density function}
\usage{
make_cdf(option = c("Omicron", "Delta", "None"))
}
\arguments{
\item{option}{SARS-CoV-2 covariant or NULL model}
}
\value{
cumulative density function for incubation period
}
\description{
Define cumulative density function. Used especially for
calculation of incubation period. Parameters for Omicron and Delta
SARS-CoV-2 strains estimated from
https://www.eurosurveillance.org/content/10.2807/1560-7917.ES.2022.27.6.2200042.
This is not a time-varying biological quantity - any changes in the
dominant variant should be modelled separately.
}
