% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_delays.R
\name{make_ecdf}
\alias{make_ecdf}
\title{Make empirical cumulative distribution function}
\usage{
make_ecdf(y, x)
}
\arguments{
\item{y}{probability}

\item{x}{vector of one or more elements from which to choose}
}
\value{
ecdf function
}
\description{
Convert a vector of cumulative probabilities into an ecdf object.
}
