% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generation_interval_density.R
\name{make_generation_interval_density}
\alias{make_generation_interval_density}
\title{Make generation interval density}
\usage{
make_generation_interval_density(gi_distribution_data)
}
\arguments{
\item{gi_distribution_data}{}
}
\description{
Make generation interval density
}
