% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_incubation_period_cdf.R
\name{make_incubation_period_cdf}
\alias{make_incubation_period_cdf}
\title{Make incubation period cumulative density function}
\usage{
make_incubation_period_cdf(
  strain = c("Omicron", "WT", "Alpha", "Beta/Gamma", "Delta")
)
}
\arguments{
\item{strain}{}
}
\value{
cumulative density function for incubation period
}
\description{
Note that in using incubation period for modelling we should
not treat it as time-varying (due to changes in the dominant variant).
Maybe through vignette or some other documentation we should make it clear
that this is not a time-varying biological quantity and the change in
dominant variant should be modelled separately)
}
