% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_class_objs.R
\name{parametric_dist_to_curve}
\alias{parametric_dist_to_curve}
\title{Create distribution object from parametric distribution values}
\usage{
parametric_dist_to_curve(dist, min_delay = NULL, max_delay = NULL)
}
\arguments{
\item{dist}{distributional package object}

\item{min_delay}{optional specification for minimum delay}

\item{max_delay}{optional specification for maxiumum delay}
}
\value{
lowerGPreff_distribution object
}
\description{
Create distribution object from parametric distribution values
}
