% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_dev.R
\name{parametric_dist_to_distribution}
\alias{parametric_dist_to_distribution}
\title{Create distribution object from parametric distribution values}
\usage{
parametric_dist_to_distribution(dist, min_delay = NULL, max_delay = NULL)
}
\arguments{
\item{dist}{distributional package object}

\item{min_delay}{optional specification for minimum delay}

\item{max_delay}{optional specification for maxiumum delay}

\item{normalise}{logical argument whether to normalise the mass output}
}
\value{
lowerGPreff_distribution object
}
\description{
Create distribution object from parametric distribution values
}
