% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_massfun.R
\name{plot.lowerGPreff_massfun}
\alias{plot.lowerGPreff_massfun}
\title{Default plot function for lowerGPreff_massfun object}
\usage{
\method{plot}{lowerGPreff_massfun}(x, y, ...)
}
\arguments{
\item{x}{x value}

\item{y}{y value}

\item{...}{additional arguments}
}
\value{
plot
}
\description{
Default plot function for lowerGPreff_massfun object
}
