% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infection_traj.R
\name{plot_infection_traj}
\alias{plot_infection_traj}
\title{Infection trajectory visualisation}
\usage{
plot_infection_traj(infection_traj, n_shown = 10, right_truncate = 14)
}
\arguments{
\item{infection_traj}{simulated trajectories}

\item{n_shown}{how many to visualise}

\item{right_truncate}{how many days to truncate at the end}
}
\value{
plot
}
\description{
Infection trajectory visualisation
}
