% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_diag_cases.R
\name{residual_diag_cases}
\alias{residual_diag_cases}
\title{scaled residual quantile diagnostic plots using DHARMa package}
\usage{
residual_diag_cases(cases_posterior, cases_true)
}
\arguments{
\item{cases_true}{}
}
\value{

}
\description{
scaled residual quantile diagnostic plots using DHARMa package
}
