% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_check.R
\name{sim_check}
\alias{sim_check}
\title{takes a greta array timeseries (can have multiple jurisdictions as columns), plot a few trajectories, either from prior or posterior distributions. This function is intended for diagnostics and troubleshooting purposes.}
\usage{
sim_check(
  greta_array,
  n_sim = 100,
  fitted_values = NULL,
  horizontal_at_one = FALSE,
  title = NULL
)
}
\arguments{
\item{greta_array}{the greta array to visualise}

\item{n_sim}{number of trajectories to plot}

\item{fitted_values}{if plotting posteriors, posterior draws}

\item{horizontal_at_one}{if a horizontal line at y = 1 should be drawn}

\item{title}{plot title}
}
\value{
plot
}
\description{
takes a greta array timeseries (can have multiple jurisdictions as columns), plot a few trajectories, either from prior or posterior distributions. This function is intended for diagnostics and troubleshooting purposes.
}
