% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dow_correction_arrays.R
\name{create_dow_correction_arrays}
\alias{create_dow_correction_arrays}
\title{Create day of week correction arrays}
\usage{
create_dow_correction_arrays(n_jurisdictions = 1, data_id = "")
}
\arguments{
\item{n_jurisdictions}{number of jurisdictions, defaults to 1}

\item{data_id}{optional name label identifying data type for greta arrays}
}
\value{
named greta arrays for day of week corrections
}
\description{
A short description...
}
