% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infection_timeseries.R
\name{create_infection_timeseries}
\alias{create_infection_timeseries}
\title{Create infection timeseries}
\usage{
create_infection_timeseries(
  n_days_infection,
  n_jurisdictions = 1,
  effect_type = c("infections", "growth_rate", "growth_rate_deriv")
)
}
\arguments{
\item{n_days_infection}{length of full date sequence of infection timeseries}

\item{n_jurisdictions}{number of jurisdictions, defaults to 1}

\item{effect_type}{options include c('infections', 'growth_rate',
'growth_rate_derivative')}
}
\value{
greta arrays for infection timeseries
}
\description{
Compute the number of infections per day, given a
log-scaled initial number of infections and a time-varying random effect
controlling the trend of infections over time. The random effect is defined
as a Gaussian Process (GP), and can be applied to the infection number
through three different formulations.

This random effect over time is not explicitly specified as an
epidemiologically defined quantity like the growth rate or the reproduction
number. Rather, these quantities are calculated from posterior samples.
The function uses an uninformative prior for the initial number of infections.
}
