% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observation_model.R
\name{create_observation_model}
\alias{create_observation_model}
\title{Create observation model}
\usage{
create_observation_model(
  infection_timeseries,
  infection_delay_distribution,
  timevarying_proportion,
  count_data,
  data_id
)
}
\arguments{
\item{data_id}{optional name label identifying data type for greta arrays}
}
\description{
Create observation model
}
