% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observation_model.R
\name{create_observation_model}
\alias{create_observation_model}
\title{Create observation model}
\usage{
create_observation_model(
  infection_timeseries,
  delay_distribution,
  proportion_observed,
  count_data,
  dow_model = NULL,
  data_id = NULL
)
}
\arguments{
\item{dow_model}{optional module of greta arrays defining day-of-week model}

\item{data_id}{optional name label identifying data type for greta arrays}
}
\description{
A short description...
}
