% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lowerGPreff_massfun.R
\name{data_to_distribution}
\alias{data_to_distribution}
\title{Create distribution object from data}
\usage{
data_to_distribution(data, min_delay = NULL, max_delay = NULL)
}
\arguments{
\item{data}{delay data}

\item{min_delay}{optional specification for minimum delay}

\item{max_delay}{optional specification for maxiumum delay}
}
\value{
lowerGPreff_distribution object
}
\description{
Create distribution object from data
}
