% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_matrix.R
\name{data_to_matrix}
\alias{data_to_matrix}
\title{Turn wide data to matrix}
\usage{
data_to_matrix(long_data, extra_col)
}
\arguments{
\item{long_data}{long data form}

\item{extra_col}{for column name}
}
\value{
matrix form
}
\description{
Turn wide data to matrix
}
