% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_observation_data.R
\name{define_observation_data}
\alias{define_observation_data}
\title{Define observation data}
\usage{
define_observation_data(
  timeseries_data,
  delay_from_infection,
  proportion_infections,
  dow_model = NULL
)
}
\arguments{
\item{timeseries_data}{timeseries data for data of interest}

\item{delay_from_infection}{delay data}

\item{proportion_infections}{proportion data}

\item{dow_model}{logical indicating whether to apply at DOW}
}
\value{
list of observation data for one data type
}
\description{
Define observation data
}
