% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_observation_model.R
\name{define_observation_model}
\alias{define_observation_model}
\title{Define observation model}
\usage{
define_observation_model(
  target_infection_dates = NULL,
  target_jurisdictions,
  data_inform_inits = "cases",
  x,
  ...
)
}
\arguments{
\item{x}{observation data set}

\item{...}{optional additional observation data sets}
}
\value{
list of datasets
}
\description{
Placeholder function for now to create list of the observation datasets.
}
