% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_waves.R
\name{fit_waves}
\alias{fit_waves}
\title{Fit waves to observations}
\usage{
fit_waves(
  observations,
  infection_model_type = c("flat_prior", "gp_infections", "gp_growth_rate",
    "gp_growth_rate_deriv"),
  n_chains = 4,
  max_convergence_tries = 3,
  warmup = 1000,
  n_samples = 2000,
  n_extra_samples = 1000
)
}
\arguments{
\item{observations}{prepared datasets}

\item{infection_model_type}{options include 'flat_prior', 'infections', 'growth_rate',
'growth_rate_derivative'. See description for more info.}

\item{n_chains}{number of chains to run MCMC}

\item{max_convergence_tries}{number of times to repeat run to get convergence}

\item{warmup}{how much samples to use in warmup}

\item{n_samples}{number of samples after warmup}

\item{n_extra_samples}{number of extra samples if original run didn't converge}

\item{target_infection_dates}{infection dates that cover more than the data dates}
}
\value{
list of infection_model, fit, infection_days, and jurisdictions
}
\description{
Compute the number of infections per day either with an
uninformative flat prior (effect_type = 'flat_prior'), or given a
log-scaled initial number of infections and a time-varying random effect
controlling the trend of infections over time. The random effect is defined
as a Gaussian Process (GP), and can be applied to the infection number
through three different formulations: effect_type = 'gp_infections',
'gp_growth_rate', or 'gp_growth_rate_derivative'.

gp_infections: the log of infections tend towards the GP mean value in the
now-cast period.
gp_growth_rate: the growth rate tends to 1 and the infections stay at
around the same level in now-casts and forecasts.
gp_growth_rate_derivative: in now-cast/forecast the infection trajectory
would follow the most recent growth rate trend.
}
