% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_day_of_week.R
\name{implement_day_of_week}
\alias{implement_day_of_week}
\title{Implement day of week effect in observation model}
\usage{
implement_day_of_week(infection_days, dow_model)
}
\arguments{
\item{dow_model}{}
}
\description{
Implement day of week effect in observation model
}
