% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits_by_jurisdiction.R
\name{inits_by_jurisdiction}
\alias{inits_by_jurisdiction}
\title{Define initials values by jurisdiction}
\usage{
inits_by_jurisdiction(
  n_juris_ID,
  obs_data,
  delays,
  obs_prop,
  target_infection_dates
)
}
\arguments{
\item{n_juris_ID}{numbered jurisdiction index}

\item{obs_data}{matrix form of data to inform infection initials}

\item{delays}{epiwave timeseries delay object}

\item{obs_prop}{single numeric value of proportion}

\item{target_infection_dates}{infection date sequence}
}
\value{
initial values as matrix
}
\description{
Define initials values by jurisdiction
}
