% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_delays.R
\name{make_ecdf}
\alias{make_ecdf}
\title{Make empirical cumulative distribution function}
\usage{
make_ecdf(y, x)
}
\arguments{
\item{x}{}
}
\description{
convert a vector fo cumulative probabilities into an ecdf object
}
