% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generation_interval_density.R
\name{make_generation_interval_density}
\alias{make_generation_interval_density}
\title{Make generation interval density}
\usage{
make_generation_interval_density(gi_distribution_data)
}
\arguments{
\item{gi_distribution_data}{dataframe with mean and sd of generation
interval for COVID-19}
}
\value{
generation interval density function
}
\description{
Make generation interval density following data from the
literature. This function requires inputs of mean and sd estimates of the
generation interval on the log scale, and outputs a GI distribution in the
bespoke format used by this package. By default this function expects
input relevant to COVID-19, specifically, samples provided by Nishiura et
al 2020 (https://doi.org/10.1016/j.ijid.2020.02.060). This estimation is
for serial interval, not generation interval, since the latter was
difficult to observe at the time of the Nishiura publication. However,
comparison against more recent household studies shows the Nishiura
estimates still accurrately reflect our current understanding of likely
COVID GI.
}
