% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_incubation_period_cdf.R
\name{make_incubation_period_cdf}
\alias{make_incubation_period_cdf}
\title{Make incubation period cumulative density function}
\usage{
make_incubation_period_cdf(strain = c("Omicron", "Delta"))
}
\arguments{
\item{strain}{}
}
\value{
cumulative density function for incubation period
}
\description{
Define cumulative density function of incubation period.
Parameters for different COVID-19 strains estimated from
https://www.eurosurveillance.org/content/10.2807/1560-7917.ES.2022.27.6.2200042.
This is not a time-varying biological quantity - any changes in the
dominant variant should be modelled separately.
}
