% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_observation_data.R
\name{prepare_observation_data}
\alias{prepare_observation_data}
\title{Prepare observation data}
\usage{
prepare_observation_data(
  observation_data,
  target_infection_dates,
  target_jurisdictions
)
}
\arguments{
\item{observation_data}{long form data of counts of notifications}

\item{target_infection_dates}{full date sequence of infection timeseries}

\item{target_jurisdictions}{jurisdictions}
}
\value{
greta arrays of observation model
}
\description{
Prepare list of data objects needed for observation model.
}
