% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_observation_data.R
\name{prepare_observation_data}
\alias{prepare_observation_data}
\title{Prepare observation data}
\usage{
prepare_observation_data(
  timeseries_data,
  delay_from_infection,
  proportion_infections,
  type = c("count", "prevalence"),
  dow_model = NULL
)
}
\arguments{
\item{timeseries_data}{long form data of counts of notifications}

\item{delay_from_infection}{distribution of delays, covering length of
infection timeseries, including incubation period, if applicable}

\item{proportion_infections}{single fixed value of expected proportions of
infections observed in count data}

\item{type}{count or prevalence}

\item{dow_model}{optional module of greta arrays defining day-of-week model}
}
\value{
greta arrays of observation model
}
\description{
Prepare list of data objects needed for observation model.
}
